Use flags for producer. Todo: do the same for consumer
