--- producer.go
@@ -1,40 +1,0 @@
-package main
-
-import (
-	"log"
-	"os"
-
-	"github.com/streadway/amqp"
-)
-
-func failOnError(err error, msg string) {
-	if err != nil {
-		log.Fatalf("%s: %s", msg, err)
-	}
-}
-
-func main() {
-	conn, err := amqp.Dial("amqp://guest:guest@localhost:5672/")
-	failOnError(err, "Failed to connect to RabbitMQ")
-	defer conn.Close()
-
-	ch, err := conn.Channel()
-	failOnError(err, "Failed to open a channel")
-	defer ch.Close()
-
-	body := os.Args[1]
-	exchange := os.Args[2]
-	routingkey := os.Args[3]
-	err = ch.Publish(
-		exchange,   // exchange
-		routingkey, // routing key
-		false,      // mandatory
-		false,      // immediate
-		amqp.Publishing{
-			ContentType: "text/plain",
-			Body:        []byte(body),
-		})
-	failOnError(err, "Failed to publish a message")
-
-	log.Printf(" [x] Sent \"%s\"", body)
-}
